backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
email-validator==2.2.0
apscheduler==3.10.4
sentence-transformers==3.1.1
numpy==1.26.4
faiss-cpu==1.8.0
transformers==4.44.2
torch==2.3.1
celery==5.4.0
redis==5.0.8
